Revert `fixed_t` from `long` to `int`.
There must be some overflows somewhere in this codebase, because
making `fixed_t` 64-bit causes `30uv1729.lmp` to desync.

+++ source/m_fixed.h
@@ -24,9 +24,9 @@
 
 /* Fixed point, 32bit as 16.16. */
 #define FRACBITS                16
-#define FRACUNIT                (1l<<FRACBITS)
+#define FRACUNIT                (1<<FRACBITS)
 
-typedef long fixed_t;
+typedef int fixed_t;
 
 fixed_t FixedMul        (fixed_t a, fixed_t b);
 fixed_t FixedDiv        (fixed_t a, fixed_t b);
